metacritic_scraper/requirements.txt
bs4
lxml
requests
pandas
progressbar
kafka-python
PyYAML
inquirer